power work

+++ src/hal/hal_timer.h
@@ -22,7 +22,7 @@
 extern "C" {
 #endif
 
-void hal_timer_init(void (*isr)(void), uint32_t interval_us);
+void hal_timer_init(void (*isr)(void), uint16_t interval_us);
 
 void hal_timer_delay(uint16_t msdelay);
